Move security-related parts of RestApiAutoConfiguration to SecurityAutoConfiguration
