+++ modules/activiti-spring-boot/spring-boot-samples/spring-boot-sample-rest-api/src/test/java/org/activiti/test/spring/boot/RestApiAutoConfigurationTest.java
@@ -5,6 +5,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.activiti.spring.boot.DataSourceProcessEngineAutoConfiguration;
 import org.activiti.spring.boot.RestApiAutoConfiguration;
+import org.activiti.spring.boot.SecurityAutoConfiguration;
 import org.junit.After;
 import org.junit.Test;
 import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
@@ -35,6 +36,7 @@ public class RestApiAutoConfigurationTest {
           ServerPropertiesAutoConfiguration.class,
           DataSourceAutoConfiguration.class,
           DataSourceProcessEngineAutoConfiguration.DataSourceProcessEngineConfiguration.class,
+          SecurityAutoConfiguration.class,
           RestApiAutoConfiguration.class,
           JacksonAutoConfiguration.class
   })
