+++ modules/activiti-spring-boot/spring-boot-starters/activiti-spring-boot-starter-basic/src/main/java/org/activiti/spring/boot/RestApiAutoConfiguration.java
@@ -17,18 +17,12 @@ package org.activiti.spring.boot;
 
 import org.activiti.rest.common.application.ContentTypeResolver;
 import org.activiti.rest.common.application.DefaultContentTypeResolver;
-import org.activiti.rest.security.BasicAuthenticationProvider;
 import org.activiti.rest.service.api.RestResponseFactory;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.servlet.configuration.EnableWebMvcSecurity;
 
 /**
  * Auto-configuration and starter for the Activiti REST APIs.
@@ -64,28 +58,4 @@ public class RestApiAutoConfiguration {
   	
   }
 
-  @Configuration
-  @ConditionalOnClass(name = {"org.activiti.rest.service.api.RestUrls", "org.springframework.web.servlet.DispatcherServlet"})
-  @EnableWebSecurity
-  @EnableWebMvcSecurity
-  public static class SecurityConfiguration extends WebSecurityConfigurerAdapter {
-    
-    @Bean
-    public AuthenticationProvider authenticationProvider() {
-      return new BasicAuthenticationProvider();
-    }
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-      http
-        .authenticationProvider(authenticationProvider())
-        .csrf().disable()
-        .authorizeRequests()
-          .anyRequest().authenticated()
-          .and()
-        .httpBasic();
-    }
-  }
-
-  
 }
